refactor: extract roadrunner-console module for shared Bars/ConsoleTheme + jline
