+++ pom.xml
@@ -383,6 +383,7 @@
         <module>roadrunner-latency-recording</module>
         <module>roadrunner-microbenchmarks</module>
         <module>roadrunner-reports-html</module>
+        <module>roadrunner-console</module>
         <module>roadrunner-reports-console</module>
         <module>roadrunner-csv-output</module>
         <module>roadrunner-parameter-source-csv</module>
