+++ roadrunner-console/src/main/java/io/roadrunner/console/ConsoleTheme.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.roadrunner.reports.console.render;
+package io.roadrunner.console;
 
 import org.jline.terminal.Terminal;
 import org.jline.utils.AttributedStringBuilder;
