+++ roadrunner-console/src/main/java/module-info.java
@@ -13,17 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import io.roadrunner.api.reports.ReportGeneratorProvider;
-import io.roadrunner.reports.console.ConsoleReportGeneratorProvider;
+module io.roadrunner.console {
+    requires transitive org.jline.terminal;
 
-module io.roadrunner.reports.console {
-    requires io.roadrunner.api;
-    requires io.roadrunner.hdrhistogram;
-    requires org.apache.commons.io;
-    requires io.roadrunner.console;
-
-    exports io.roadrunner.reports.console;
-
-    provides ReportGeneratorProvider with
-            ConsoleReportGeneratorProvider;
+    exports io.roadrunner.console;
 }
