+++ roadrunner-reports-console/pom.xml
@@ -25,8 +25,9 @@
             <artifactId>commons-io</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.jline</groupId>
-            <artifactId>jline-terminal</artifactId>
+            <groupId>io.roadrunner</groupId>
+            <artifactId>roadrunner-console</artifactId>
+            <version>${project.version}</version>
         </dependency>
 
         <dependency>
