+++ roadrunner-reports-console/src/main/java/io/roadrunner/reports/console/render/ReportRenderer.java
@@ -15,6 +15,7 @@
  */
 package io.roadrunner.reports.console.render;
 
+import io.roadrunner.console.ConsoleTheme;
 import io.roadrunner.reports.console.ReportModel;
 import java.util.Locale;
